# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "Smm Api Documentation",
        "contact": {
            "email": "takahiroasiro@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/articles/{id}": {
            "put": {
                "tags": [
                    "Article"
                ],
                "summary": "update articles",
                "description": "Обновление статей",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "request",
                        "in": "query",
                        "description": "The request object parameter in query",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Article"
                ],
                "summary": "destroy articles",
                "description": "Удаление статей",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The request id parameter in query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles deleted successfully"
                    }
                }
            }
        },
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get list of notifications",
                "description": "Returns list of notifications",
                "operationId": "getNotifications",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Post a new notification",
                "description": "Returns list of notifications",
                "operationId": "postNotifications",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotificationPostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/notifications/{id}": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Update selected notification",
                "description": "Returns updated notification",
                "operationId": "getNotifications",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotificationPostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad data was given."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Delete selected notification",
                "description": "Delete selected notification",
                "operationId": "destroyNotifications",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/admin/users/status": {
            "post": {
                "tags": [
                    "PremiumStatus"
                ],
                "summary": "Update PremiumStatus",
                "description": "Изменение статуса в программе лояльности",
                "operationId": "updateBalance",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The id parameter in query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "The status parameter in query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/userjobs/{id}": {
            "post": {
                "tags": [
                    "UserJobs"
                ],
                "summary": "Create new action",
                "description": "Создает action для локального задания",
                "operationId": "create",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The id parametr in query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok, action created"
                    }
                }
            }
        },
        "/api/userjobs/check/{actionId}": {
            "post": {
                "tags": [
                    "UserJobs"
                ],
                "summary": "Check user job status",
                "description": " Запрос на проверку статуса задания",
                "operationId": "check",
                "parameters": [
                    {
                        "name": "actionId",
                        "in": "query",
                        "description": "The actionId parametr in query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат поиска"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "NotificationPostRequest": {
                "title": "Creating request for a notification",
                "description": "Creating request for a notification",
                "properties": {
                    "content": {
                        "title": "Content",
                        "description": "Notification body",
                        "format": "string",
                        "example": "We have a lunch today"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
    "security": [
        []
    ]
}